requirements.txt
absl-py==0.11.0
apache-beam==2.28.0
args==0.1.0
astunparse==1.6.3
attrs==20.3.0
avro-python3==1.10.1
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clint==0.5.1
colorama==0.4.4
comtypes==1.1.10
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
dill==0.3.1.1
dm-tree==0.1.5
docopt==0.6.2
docutils==0.16
fastavro==1.3.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.1.2
gin-config==0.4.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.32.0
h5py==2.10.0
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
importlib-resources==5.1.1
joblib==1.0.1
Js2Py==0.71
kaggle==1.5.10
Keras-Preprocessing==1.1.2
keyring==22.3.0
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.4
mock==2.0.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pbr==5.5.1
Pillow==8.1.0
pipwin==0.5.1
pkginfo==1.7.0
promise==2.3
proto-plus==1.14.2
protobuf==3.15.3
psutil==5.8.0
py-cpuinfo==7.0.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.8.0
pyjsparser==2.7.1
pymongo==3.11.3
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pySmartDL==1.3.4
python-dateutil==2.8.1
python-slugify==4.0.1
pyttsx3==2.90
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-metadata==0.28.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tqdm==4.58.0
twine==3.3.0
typeguard==2.11.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
